+++ ChangeLog.md
@@ -1,6 +1,9 @@
 ## change log ##
 
 * 2016-06-20: 3.0.alpha:
+  - implement #15 "v2r: for column property definition allow selection from standard vocabularies"
+    A selected property is displayed with associated typical prefix, eg., skos:definition (#19).
+    The list is for now hard-coded. TODO capture list in a registered (internal) ontology.
   - implement #15 "branding"
 
 * 2016-06-18: 0.3.2: